compilation fixes for VS2010 on win7 x64

+++ src/name_list.h
@@ -8,7 +8,7 @@ struct NameList
 {
     std::vector<std::string> names;
 
-    int count() const { return names.size(); }
+    int count() const { return (int) names.size(); }
     const char* operator[](int index) const { return names[index].c_str(); }
 };
